# Conflicts:
#	domenetjenester/arbeidsforhold/src/test/java/no/nav/ung/sak/domene/abakus/mapping/IAYDtoMapperRoundtripTest.java

+++ formidling/src/main/java/no/nav/ung/sak/formidling/innhold/GrunnlagOgTilkjentYtelse.java
@@ -1,15 +1,15 @@
 package no.nav.ung.sak.formidling.innhold;
 
-import java.math.BigDecimal;
-
 import no.nav.ung.kodeverk.ungdomsytelse.sats.UngdomsytelseSatsType;
+
+import java.math.BigDecimal;
 
 /**
  * Intermediary objekt for å periodisere felter for beregning og tilkjent ytelse i brev.
  */
 public record GrunnlagOgTilkjentYtelse(
     long dagsats,
-    BigDecimal utbetalingsgrad,
+    long utbetalingsgrad,
     UngdomsytelseSatsType satsType,
     BigDecimal grunnbeløpFaktor,
     long grunnbeløp,
